Move legacy block to “legacy” folder
