+++ assets/js/legacy/views/attribute-select.jsx
