+++ assets/js/legacy/views/category-select.jsx
